refactor: rename user_banner_picker -> user_banner_picked
