+++ lib/app/features/user/pages/profile_edit_page/profile_edit_page.dart
@@ -20,7 +20,7 @@ import 'package:ion/app/features/user/pages/components/profile_avatar/profile_av
 import 'package:ion/app/features/user/pages/profile_edit_page/components/category_selector/category_selector.dart';
 import 'package:ion/app/features/user/pages/profile_edit_page/components/edit_submit_button/edit_submit_button.dart';
 import 'package:ion/app/features/user/pages/profile_edit_page/components/header/header.dart';
-import 'package:ion/app/features/user/pages/profile_edit_page/components/user_banner_picked/user_banner_picker.dart';
+import 'package:ion/app/features/user/pages/profile_edit_page/components/user_banner_picked/user_banner_picked.dart';
 import 'package:ion/app/features/user/pages/profile_edit_page/hooks/use_draft_metadata.dart';
 import 'package:ion/app/features/user/providers/user_metadata_provider.dart';
 import 'package:ion/app/router/app_routes.dart';
